move: 공용 컴포넌트 이동 및 폴더 구조 정리
