+++ src/pages/landing/components/LandingPC/components/ThirdSection.tsx
@@ -6,7 +6,7 @@ import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import BookshelfCard from '../../BookshelfCard';
 import LandingButton from '../../LandingButton';
-import ShelfDecoration from '../../ShelfDecoration';
+import ShelfDecoration from '../../../../../components/shelfDecoration/ShelfDecoration';
 import { Main as BaseMain, Layout as BaseLayout, ButtonContainer } from '../style/commonPC';
 
 interface TThirdSection {
